+++ packages/moonstone/CHANGELOG.md
@@ -4,9 +4,14 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 ## [unreleased]
 
+### Added
+
+- `moonstone/Scrollable` scroll to top button
+
 ### Fixed
 
-- `moonstone/Notification`to support 3 max-width buttons in a single line
+- `moonstone/Notification` to support 3 max-width buttons in a single line
+
 ## [3.2.5] - 2019-11-14
 
 ### Fixed