+++ packages/moonstone/CHANGELOG.md
@@ -7,6 +7,11 @@ The following is a curated list of changes in the Enact moonstone module, newest
 ### Fixed
 
 - `moonstone/Notification` to support 3 max-width buttons in a single line
+## [unreleased]
+
+### Fixed
+
+- `moonstone/Notification`to support 3 max-width buttons in a single line
 
 ## [3.2.4] - 2019-11-07