+++ packages/moonstone/CHANGELOG.md
@@ -8,6 +8,9 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 - `moonstone/Notification` to support 3 max-width buttons in a single line
 ## [unreleased]
+- `moonstone/VirtualList.VirtualList` to render properly without error when `itemSizes` is given and `dataSize` is 0
+
+## [3.2.5] - 2019-11-14
 
 ### Fixed